Starting to test sensor fusion with madgwick

+++ src/sensor_fusion.c
@@ -3,10 +3,9 @@
 #include "math.h"
 #include "inttypes.h"
 
-int32_t calculateRoll(SensorData *d) {
-    int32_t roll;
+double calculateRoll(SensorData *d) {
+    double roll;
 
     roll = atan2(d->acc[X], d->acc[Y]);
     return roll;
-
 }
